Cart is rendering from database

+++ db/products_table_seed.sql
@@ -1,7 +1,7 @@
 create table products(
     id serial primary key,
     product text,
-    price integer,
+    price decimal(4,2),
     category varchar(20),
     image_src text
 )
